nearly done! :o

+++ src/main/java/com/yachtmafia/db/DBWrapper.java
@@ -2,13 +2,14 @@ package com.yachtmafia.db;
 
 import com.yachtmafia.messages.SwapMessage;
 
+import java.math.BigDecimal;
 import java.sql.SQLException;
 
 public interface DBWrapper {
     String getPublicAddress(String user, String coin) throws SQLException;
     boolean addNewWallet(String user, String coin, String publicAddress, String privateAddress);
     boolean addPortfolioBalance(SwapMessage message, String purchasedAmount);
-    Double getFunds(String user, String coin);
+    BigDecimal getFunds(String user, String coin);
     String getPrivateKey(String user, String coin);
 
 }
