+++ src/main/java/com/yachtmafia/db/DBWrapperMock.java
@@ -3,6 +3,8 @@ package com.yachtmafia.db;
 import com.yachtmafia.cryptoKeyPairs.BTC;
 import com.yachtmafia.messages.SwapMessage;
 
+import java.math.BigDecimal;
+
 public class DBWrapperMock implements DBWrapper {
     private BTC btc = new BTC();
 
@@ -25,8 +27,8 @@ public class DBWrapperMock implements DBWrapper {
     }
 
     @Override
-    public Double getFunds(String user, String coin) {
-        return 1d;
+    public BigDecimal getFunds(String user, String coin) {
+        return BigDecimal.valueOf(1d);
     }
 
     @Override
